feat: implemented dedicated conversation chat view with routing updatse
- created ConversationView.vue for separation of concerns and better
maintainability
- preserve all features
- refactor Messaginv.vue to handle conversation list oly
-
